fixing some import dependencies
